sdk/pyproject.toml
[project]
name = "re-client"
version = "0.1.0"
description = "Python SDK for Mateus-Lacerda Reranker Server"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Pacto", email = "dev@pacto.com"}
]
license = {text = "MIT"}
keywords = ["colbert", "reranking", "nlp", "grpc", "sdk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "grpcio>=1.50.0",
    "protobuf>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/pacto/re-server"
Repository = "https://github.com/pacto/re-server"
Documentation = "https://github.com/pacto/re-server/tree/main/sdk"
Issues = "https://github.com/pacto/re-server/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["re_client"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
